Update hello_world.go
Download and install the Go tools on your machine. Write a Go program to print “Hello, world!” on the screen. Compile and run the program.

Submit a screenshot of your cmd window showing you compiling and executing your code. The string “Hello, world!” should appear on the screen when you execute the code.

hello_world